requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
cloudpickle==3.0.0
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
dgl==0.7.1
fair-esm==2.0.0
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
future==1.0.0
hyperopt==0.2.7
idna==3.7
importlib_resources==6.4.0
Jinja2==3.1.3
joblib==1.4.0
kiwisolver==1.4.5
lmdb==1.4.1
MarkupSafe==2.1.5
matplotlib==3.7.5
multidict==6.0.5
networkx==3.1
ninja==1.11.1.1
numpy==1.22.4
ordered-set==4.1.0
packaging==24.0
pandas==2.0.3
pillow==10.3.0
psutil==5.9.8
py4j==0.10.9.7
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
rdkit-pypi==2022.9.5
requests==2.31.0
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
threadpoolctl==3.4.0
torch_geometric==2.5.3
torchdrug==0.2.1
tqdm==4.66.2
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4
zipp==3.18.1
